Add form input validation for LoginForm.tsx
